Always pull Docker image
Existing behaviour is to only pull DOCKER_IMAGE when
using the image as is. If the DOCKER_FILE path is
traversed, the base DOCKER_IMAGE is not refreshed
if it exists on the system. Fix is to move the pull
before the 'if' so that it is run for both cases.

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>