Update build status to count unstable builds as complete
A change to the build jobs has resulted in the status of all
jobs downstream from the build job (mostly test jobs) affecting
the status of the build job itself.

This is helpful in many ways, however it does make it harder for the
build status report tool to identify an actual build failure, as
opposed to a failure in a subjob.

To reduce the impact in the short term, I'm telling the build status
tool to only report a build as failed if it explicitly has the FAILED
status. That way all the "UNSTABLE" test jobs can't fool us into
an exaggerated build failure rate when its status is inherited by
the parent build job.

Signed-off-by: Adam Farley <adfarley@redhat.com>