This allows us to run unpublished builds with the weekly test set
In the long run we should find a neater way to run ea
specific-tagged-level build testing. Perhaps when all of Adoptium's
builds are ea (and nightly/weekly becomes an unused concept), we
could have a single test list for all builds.

This seems a good solution in the short term though.

Signed-off-by: Adam Farley <adfarley@redhat.com>