Removed the endless growth of path
Also put the "expected" java into the start, as otherwise the system JDK
would be always before the "expected" jdk. I think the meddling wiht
PATH is actually useless, but I do not see to the bottom of this script.
Am in temptation to remove that at all. But that can bite