Making jdk for jtreg build explicit (#1019)
* Making jdk for jtreg build explicit

and so imune against change of system jdk on underlying system/image

* Removed the endless growth of path

Also put the "expected" java into the start, as otherwise the system JDK
would be always before the "expected" jdk. I think the meddling wiht
PATH is actually useless, but I do not see to the bottom of this script.
Am in temptation to remove that at all. But that can bite

* Removed duplicated build of jtreg 6_1

---------

Co-authored-by: Martijn Verburg <martijnverburg@gmail.com>