avoids image name error when pulling from non-default registery
when pulling from a non-default registery, you will have local image in full name include docker registery in it. But the whole pipeline works with DEFAULT behaviour when pulling from dockerhub that creates local image with image name only. This change, adds a tag to pulled local image with image_name only and avoids issue in continue. Also as a result of changes we do not need to clean `dockerImageDigest` later in the code. So that line removed.

Signed-off-by: mahdi@ibm.com