Making sure we use the latest build repro info during healthcheck
Currently, we use the earliest pipeline with a given tag as our
best guess at the root source of a published build.

However, if the build link is rerun by a user for specific
platforms, the original link becomes outdated, and we should
use the new link for healthchecks (but only on the affected
platform).

This change adds the per-platform logic required to do that. We may
expand the scope for this in the long run, but for now this logic
will only be used by the section of the code which identifies the
reproducibility percentages for published builds.

Signed-off-by: Adam Farley <adfarley@redhat.com>